backend/pyproject.toml
[project]
name = "studyhub-backend"
version = "0.1.0"
description = "StudyHub FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pyjwt>=2.10.0",
    "pwdlib[argon2]>=0.2.0",
    "litellm>=1.55.0",
    "valkey>=6.0.0",
    "slowapi>=0.1.9",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.1.0",
    "httpx>=0.28.0",
    "boto3>=1.36.0",
    "temporalio>=1.9.0",
    "pyalex>=0.21",
    "rapidfuzz>=3.14.3",
    "tenacity>=9.1.4",
    "python-multipart>=0.0.22",
    "meilisearch-python-sdk>=7.0.4",
    "patchright>=1.49.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "neo4j>=5.27.0",
    "scholarly>=1.7.0",
    "jinja2>=3.1.6",
    "huggingface-hub>=1.7.1",
    "scikit-learn>=1.8.0",
    "umap-learn>=0.5.11",
    "gitpython>=3.1.0",
    "docker>=7.0.0",
    "nvidia-ml-py>=12.0.0",
    "weasyprint>=68.1",
    "markdown>=3.10.2",
    "matplotlib>=3.10.8",
    "pillow>=12.1.1",
    "pymupdf>=1.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
